feat: добавляет кнопку профиля
