+++ src/components/header/header.tsx
@@ -6,10 +6,10 @@ import {store} from '../../store';
 import {useAppSelector} from '../../hooks';
 import {UserData} from '../../types/user';
 import {selectAuthStatus, selectUserData} from '../../store/user-process/selectors';
+import ProfileLink from '../profile-link/profile-link';
 
 function Header():ReactNode {
   const authorizationStatus: AuthorizationStatus = useAppSelector(selectAuthStatus);
-  const user: UserData = useAppSelector(selectUserData);
 
   const handleLogoutClick = (event:React.MouseEvent<HTMLLinkElement>) => {
     event.preventDefault();
@@ -38,15 +38,7 @@ function Header():ReactNode {
                 </li>
                 :
                 <>
-                  <li className='header__nav-item user'>
-                    <Link className="header__nav-link header__nav-link--profile" to={AppRoute.favorites}>
-                      <div className="header__avatar-wrapper user__avatar-wrapper">
-                        <img src={user.avatarUrl} alt={user.name} />
-                      </div>
-                      <span className="header__user-name user__name">{user.email}</span>
-                      <span className="header__favorite-count">3</span>
-                    </Link>
-                  </li>
+                  <ProfileLink />
                   <li className="header__nav-item">
                     <a
                       className="header__nav-link"
